Allow Listening to Specific Hosts
This patch allows to listen to a specific host only.

For example, when BigBlueButton is used behind Nginx as reverse Proxy,
it is completely sufficient to listen to 127.0.0.1 only, but right now
there is now easy way to configure that.

Note that this patch does not change the default and still works with
the previous configuration making updates easy.