[audio] Improve source audio cleanup check to avoid wrongly stopping it
There was a weird side effect that might cause global audio bridge to be wrongfully closed when only two endpoints were connected to it and listener stop requests glared. That was because the remaining endpoint check (which checks if the listener being stopped is the only one) did not check if the remaining endpoint was itself. That, coupled with two glared requests, could trigger a source audio stop that should not be enacted.

The end user behaviour would be that ALL two listen only users would be ejected instead of just the one that requested stop (but requested it with glare)

Also improve/add additional logs to catch source audio stops and glares