kurento: disable mDNS candidates by default
Commit 377435c (v2.4.23) enabled mDNS processing by default.
This was a decision I made myself after testing it in a simulated intranet AND load testing it in regular setups.

Seeing that it worked with no apparent drawbacks, I enabled them. It seems, though, that enabling it causes stalls in Kurento
processing pipelines earlier than expected. That effectively reduces Kurento's capacity by a fair margin due to something related 
to unresolvable DNS queries stalling the reduced pool of threads Kurento uses for such stuff.

We've only seen that happen in one specific cloud provider. It's not generalized. But I'm reverting it to disabled by default until we
figure out what's up.