kurento: add a mitigation to circumvent potential KMS UUID collisions across different KMS processes
KMS uses UUIDs (v4) to identify its media pipelines and media elements. I was counting on them being unique to simply ID management in Kurento´s adapter. Turns out that was a bad call: we have seen cases where separate KMS instances spawned to close in time to one another were generating a seed that was similar or weak enough so that it caused abnormal UUID collisions.

That made the experimental stream transposing/offloading system in bbb-webrtc-sfu SFU kick in and mix media elements belonging to different pipelines, which could, rarely, cross streams between different meetings

So I added an enrichment phase that adds a host marker to stream ids internally so that we can differentiate streams from different hosts even if they have the same ID. In the case where a collision happens in a single KMS instance, I added an error check that will reject negotiation for the offender stream.

This is, of course, a mitigation until we fix the weak generator seed in KMSs uuid generator