refactor(mediasoup): re-organize server-as-offerer procedures
When we should be producing something into mediasoup, do not create one upfront;
prefer manually generating an offer based on desired capabilities and just
create the producer when the remote offer comes in.

When we should be consuming something from mediasoup, generate the consumer
beforehand based on the source's caps. When the remote offer comes, do not
create a second (useless) consumer; use the one created for the offer.

This should address a couple of problems: ssrc mismatches, consumer and
producer leaks.