feat(mediasoup): add new multi-transport RTP element
Split transport is a mediasoup-specific adapter option which
means the RTP element will have one transport per media type, meaning
that an element in our mediasoup adapter might have more than one
transport in it.

This is necessary because, otherwise, mediasoup will bundle
streams in the same port in the plain transport. This is OK,
but neither FS nor KMS support that properly, and neither do the
majority of legacy VoIP endpoins out there.

The default behavior is still bundling.