chore: keep the new full audio module disabled by default
This version is going to land on envs with 2.2, 2.3, 2.4 and 2.5; this is only
needed on 2.5 so what we want is to keep this disabled by default and then make
packaging enable it on its own in 2.5. The rationale is to not keep an idle
entrypoint wide open to the clients.

When its deemed necessary, we can alter the default here.