fix(mediasoup): fix worker replacement routine
The mediasoup worker replacement routine (which was executed in case of
worker crashes) wasn't doing the bare minimum: increment the Prometheus
crash counter, remove the crashed worker reference and replace the
worker (in the correct worker pool).

This is due to a regression when implementing the dedicatedMediaType
worker pools - the crash exporter wasn't updated with the 'pool' label
which made the replacement routine fail silently.