fix(core): set local descriptor after remote counterpart
Fixed an issue where, in glare scenarios, the negotiation role would be
flagged as OFFERER incorrectly.

The whole things sucks all around, but the tldr rationale is: remote
descriptors *may* be undefined at media unit creation time, but local ones
*will not be undefined* if a remote descriptor is unset - so local
descriptors must be processed first if they exist (for scenarios where both
local and remote are provided at the same time). That guarantees the
negotiation role will be properly flagged as ANSWERER in this specific
scenario.