fix(mediasoup): properly inherit router settings on transceiver elements
Transceiver elements are using the base router settings
(default|etc.yml) to generate offers when we're acting as the offerer.
That's wrong as it may generate payload type asymmetry in case the
default router config is different from the requesting endpoint. An
example of that is the Firefox microphone bridge between mediasoup and
FS generating offers-answers with a 109 OPUS PT for mediasoup -> FS and
a 111 OPUS PT for FS -> mediasoup. That's not what I want to see.

Now we properly inherit router settings from the transport set's router
when generating offers with no remote SDP to work with.