fix(kurento): deep clone kurento config because kurento-client mutates it
kurento-client mutates its options to add default parameters for reasons unknown to mankind

That exposed a bug after node-config was bumped to 3.3.9 (from 3.3.6) and started throwing errors if configurations are mutated without a ALLOW_CONFIG_MUTATIONS env var set

We do mutate some other configs directly, but I made sure that those are always deep cloned. In the case, it was an indirect mutation in a 3rd party library.