fix: improve recording timestamp generation
There's a problem right now where the processing scripts rely on the start rec
evt timestamp to be an approximate of the first keyframe timestamp, but the
conditional capture mode borks that behavior. That causes the final recording
to have A/V desynchronization.

Adjust recording timestamp generation when conditional capture is active - this
should improve A/V sync without needing to change RAP behavior