fix: do not clog request queues with unhandled recording exception
There's a race condition in the recording startup code when
recordFullDurationMedia=false that may cause an unhandled exception to
be thrown. That exception causes the inbound request queues to clog,
which translates to screen sharing/video randomly stopping working until
the queues are unclogged or the application is restarted.

The exception is thrown whenever the filename is still not populated in
memory when the events.xml event generation code kicks in - which is an
oversight. Furthermore, the event generation method *should not throw*
when the filename isn't present - and instead just skip and wait for the
next call to it in a later step (ie flow -> start).

This commit does a bunch of things to fix+track this edge case:
  - Guarantee the filename is always present whenever calling the event
    generation methods in places where the recording startup might have
    not fully finished yet (ie media flow callback handlers)
  - Do not throw errors when filename is not present in the event
    generation method - instead log a warning and skip
  - Add the screen share presenter's user id to the log metadata