fix: set appropriate initial bitrates
mediasoup's initial outgoing bitrate for transports are hardcoded and
won't follow what's specified by sources. This is mostly minor, but
introduces a slight delay in the estimation ramp up process.

This commit guarantees mediasoup's values are set based on client
provided profiles in screen sharing and webcams, as well as correct
Kurento's initial REMB values in a similar way.