feat: base-manager may opt out of connection to mcs-core
Subprocesses may not need to connect to mcs-core. The recent example if
LiveKit's process, which won't use the core subsystem. While this
doesn't have any visible impact, it's an extra unused WebSocket conn
which is a small waste of resources.

Add a constructor option to base-manager (connectToMCS) so that
inheritors may opt out of connection to mcs-core.