feat(livekit): add support for SIP trunking
SIP trunking/dial in integration is currently done by FreeSWITCH in BBB.
The alternative in LiveKit is their SIP trunking APIs and application,
but trunking and dispatch rules need to be implemented on our side.

Add support for LiveKit's SIP trunking API. This can be enabled via
`livekit.sip.enabled` and is off by default. The way this currently
works is:
  - Create a single SIP trunk register in LiveKit that works for the
    calling BBB node
    - The trunk's auth creds can be configured via
      livekit.sip.inbound.auth_password|auth_username
  - For each created room with a valid dial-in number, register a direct
    dispatch rule which has the voiceBridge as its PIN. That way, an
    user will need to call the dial-in number and then enter the voice
    conf PIN to land into the desired room. That is on par with the
    current production behavior.

This needs the livekit/sip interop application to be installed and
running to have any effect.