feat: add `enabled` flag to SFU modules, default to true
There's currently no way to control whether to boot up SFU subprocesses
other than removing them from the modules array in config. This is not
problematic per se, but having an "enabled" flag could make it simpler
to alter/test and deploy.

Add support for a `enabled` flag in a submodule config spec. The default
behavior is true. If false, the subprocess won't be forked at all.