fix: reorganize module forking to work around FIFO sched issues
On Node.js >= 20, the application may get stuck at 100% CPU
usage on boot whenever its under the FIFO CPU scheduling policy.
This has been isolated down to the node-redis library. tl;dr:
prototype overriding on forked processes + FIFO is causing this, but
it's most likely a Node.js bug rather than the dep's.

While the source issue hasn't been fixed, module spawning (i.e.:
process forking) has been re-organized to work around that issue:
  - Make module spawning sequential and synchronous. Start them
    one-by-one, failures means an application boot failure.
  - Add a configurable minimum delay between process forks. Default: 1s.
    See moduleSpawnDelayMs.
  - Change module start triggers to something realistic (i.e.:
    subprocess notifying that they're ready via IPC)

The trade-off is that the boot sequence now takes longer. While this is
a workaround, most of the changes are good nonetheless. Once the Node.js
issue is fixed, the likely outcome is to reduce the spawn delay to 0.