fix(livekit): guarantee uniqueness of "external" LK users
Guarantee that LK user IDs are mapped to BBB user IDs in an unique way -
if they're external (e.g.: dial-in), use SID and do the identity-to-SID
mapping internally. If they're web users, then identity is already an
unique ID.

This fixes a few inconsistencies with dial-in user states not being
computed correctly due to the user's identity being used as BBB's intId,
and the SID as the voiceUserId. Now, everything is SID for dial-ins.