refactor(livekit): cleanup recordings directly on meeting end
Cleanup active recordings for a meeting directly on meeting end. This
makes the process more orderly as we won't depend on a
recorder-initiated stop due to the LK room being deleted.

Additionally, make the recorder stop/stopAllFromMeeting synchronous so
that the room is only deleted/application is only shutdown when they
have been executed.