feat(livekit): DTMF handling
By default, the FreeSWITCH setup has multiple actions assigned to
different tones. We're not able to implement all of them with LiveKit,
but mute and deaf-mute are must-have actions.

Add a DTMF handling system to LiveKit-based BBB meetings. This is
controlled through the `livekit.sip.dtmfActions` configuration,
which is a key-value map of actions to be performed when a DTMF
sequence is received. Currently supported actions: 'toggleMute'.

Format:
```
livekit.sip.dtmfActions: {
  ActionName<String>: DTMFSequence<String> | DTMFSequence<String>[]
}
```

By default, toggleMute is assigned to digits `0` and `*`. Deaf-mute will
be implemented in subsequent patches.