fix(livekit): SIP users have hardcoded `Phone ` prefix in their names
LiveKit appends a `Phone ` prefix to SIP participant names. This happens at
livekit/protocol in a hardcoded fashion. That prefix is not ideal as it's not
localized and is redundant with BBB's own way of flagging participants
as dial-in.

Remove the hardcoded `Phone ` from the user name prior to dispatching a
participant joined event to BBB. Either use the `sip.phoneNumber` attribute
(when number isn't masked), or the participant's name/identity with the
prefix stripped manually.