fix(livekit): unhandled EjectUserFromVoiceConfSysMsg
EjectUserFromVoiceConfSysMsg is fired by BBB to eject voice users from
FS-based voice conferences.
The LiveKit controller should handle it as well - for all purposes, this
is a request to remove a participant from a room altogether. Uniquely
important for SIP endpoints as this is the only ejection pathway that
works for those.

Add handling for EjectUserFromVoiceConfSysMsg in the LiveKit controller,
removing a participant associated to the voiceUserId from the specified
room.