feat(livekit): RTC-based event source (alternative to webhooks)
bbb-webrtc-sfu uses LiveKit webhooks to synchronize and enforce state
between BBB and LiveKit. This works, but introduces a pain point when
running LiveKit nodes outside of BBB as webhooks endpoints are
statically configured per-LK node.

Connecting to LiveKit rooms through a hidden, dormant RTC agent provides
events compatible with most of the data we need. The exception is egress,
which we're not interested on. The other benefit of an RTC agent is that
it's tied to the appropriate LK node automatically (hence circumventing
the webhook limitation outlined above).

Introduce a RTC-based event source for LK in bbb-webrtc-sfu. This new
event source uses LK's Node.js SDK and is compatible with the data we
pull from webhooks.
The event source to be used is configurable via `livekit.eventSource`.
Possible values are: `webhook`, `agent`. Default is preserved
(`webhook`).

A significant refactor is included so that we support configurable event
sources without too much code duplication. An event normalization layer
is included to that end - and it'll be extended in the future for more
strict event validation.