fix(livekit): forcefully unpublish screenshare when server ejects it
There's client side code to unpublish screen streams when the server
ejects it, as well as BBB-specific business logic to clean it up there.
The LiveKit server-side doesn't have that enforcement yet, though.

Forcefully unpublish screen share streams for an user when the server
requests it. This is done via a user permission toggle due to the lack
of a server-side unpublish API.