fix(livekit): discard camera/screen events if meeting is not using LK
Audio events coming from BBB are discarded if a meeting does not have LK
as its audio bridge. The same is not done for cameras/screen share.
While this doesn't break anything, it's a waste of resources and
generates unnecessary error logs.

Ignore camera/screen events from BBB in the LK module if a meeting is not
using LiveKit. Additionally, refactor how LK usage check is done so that
it's streamlined and re-usable.