feat: getRecordings primitive in bbb-webrtc-recorder client
Implement the getRecordings primitive add in bbb-webrtc-recorder@v0.13.0
in the bbb-webrtc-recorder client. See bbb-webrtc-recorder@v0.13.0
README for a spec on that new API call.

Additionally: add a new optional flag to the recorder client to specify
whether it should notify recordings as stopped or not on a crash.
This is going to be used for an experimental resync feature with the
LiveKit module.