feat(livekit): add recording resync
There's currently no recording procedure in place when using
bbb-webrtc-recorder for either the mediasoup or LiveKit adapter. This is
problematic in cases where heartbeats are lost on overloaded servers or
some other issue causes the recorder conn to be lost momentarily. E.g.:
  - State becomes desynced
  - Recordings might be aborted unnecessarily if they are still active
    in the recorder at reconn time

Implement recording resync in the LiveKit module using the newly added
getRecordings API from bbb-webrtc-recorder. This is currently restricted
to LiveKit as to not cause any unforeseen issues in the default/prod
setup (mediasoup). This is also the reason why some logic that could be
inside the base bbb-webrtc-recorder.js wrapper (e.g.: orphaned session
management, stop events based on resync) is isolated in the livekit
recording manager instead.

Once this proves stable, the basic resync logic will be moved to the
bbb-webrtc-recorder base wrapper.