fix(livekit): BBB state sync stalls on idle rooms
When using the "agent" eventSource mode, state sync may stall when a
room is idle only in LK. This is caused by the departure timeout
kicking in for LK, but not for BBB - hence causing a room (and event
agent) cleanup.

Alter the synchronization RTC client to not use the agent: true token
grant. It isn't really an agent, and flagging it as such makes it not
count towards room activity and causes the departure timeout to kick
in. This should guarantee the sync client stays up until the BBB meeting
ends.