fix(livekit): add conn retry logic for RTC state sync agents
RTC-based event sync clients do not handle disconnects/failures
properly. That means restarting a LiveKit node for an ongoing meeting
breaks state synchronization when eventSource == agent. This is not an
issue in the webhooks event source, for example.

Add proper disconnection handling to the RTC-based event source agents
with a retry logic similar to the one used for bbb-webrtc-recorder. This
applies to both startup failures and established conn failures.
This includes retry metrics so that this can be observable.

What we should do when retries and startup failures happens is still up
to debate. The max amount of retries is also significantly large to the
point that it does not make much sense, but it should be reduced once we
figure out what to do in the "catastrophic" failure scenario.