fix(livekit): race condition on token generation (clients fail to connect)
The token generation procedure has a race condition where it might be
processed _before_ the RoomServiceClient has finished fully setting up
the room. If it isn't present, an unhandled exception is thrown. For the
end user, this manifests as a connection timeout due to the lack of a
valid LiveKit token.
We need the room to be setup so that its metadata is available for token
generation.

This change makes the token generation procedure synchronous in relation
to the room setup. There's also timeout in place so that token requests
aren't stuck due to the RSC not resolving - _if_ that happens, a basic
set of participant metadata is used to generate the token. While
incomplete, it should guarantee a functional session for the affected
user.

Additionally, add a new Prometheus metric to track token generation
failures: `sfu_livekit_token_generation_errors_total{errorMessage}`.